+++ src/Layer/include/Core/TextNodePrivate.h
@@ -2,8 +2,8 @@
 #include "Core/FontManager.h"
 #include "Core/TextNode.h"
 #include "Core/VType.h"
-#include "SkiaBackend/SkiaImpl.h"
-#include "SkiaBackend/SkFontMgrVGG.h"
+#include "SkiaImpl/VSkia.h"
+#include "SkiaImpl/VSkFontMgr.h"
 #include "Common/DebugCanvas.h"
 #include "Core/ParagraphParser.h"
 #include <core/SkColor.h>
