+++ src/Layer/interface/Core/ContourNode.h
@@ -3,7 +3,7 @@
 #include "Core/VType.h"
 #include "Core/PaintNode.h"
 #include "Attrs.h"
-#include "SkiaBackend/SkiaImpl.h"
+#include "SkiaImpl/VSkia.h"
 
 #include <vector>
