+++ src/Layer/src/Core/PathNodePrivate.cpp
@@ -14,7 +14,7 @@
 #include "include/core/SkSurface.h"
 #include "include/core/SkTileMode.h"
 #include "include/core/SkTypes.h"
-#include "SkiaBackend/SkMyImageFilters.h"
+#include "SkiaImpl/VSkImageFilters.h"
 #include "include/effects/SkRuntimeEffect.h"
 #include "include/core/SkColor.h"
 #include "include/core/SkPath.h"
