+++ src/Layer/src/Core/TextNode.cpp
@@ -4,7 +4,7 @@
 #include "Core/Node.hpp"
 #include "Core/PaintNode.h"
 #include "Core/VType.h"
-#include "SkiaBackend/SkFontMgrVGG.h"
+#include "SkiaImpl/VSkFontMgr.h"
 #include "include/core/SkCanvas.h"
 #include "include/core/SkFont.h"
 #include "include/core/SkFontMetrics.h"
