+++ src/Layer/src/Core/TextNodePrivate.cpp
@@ -3,7 +3,7 @@
 #include "Core/ParagraphParser.h"
 #include "Core/TextNode.h"
 #include "Core/VType.h"
-#include "SkiaBackend/SkiaImpl.h"
+#include "SkiaImpl/VSkia.h"
 #include "glm/detail/qualifier.hpp"
 
 #include <core/SkCanvas.h>
